+++ src/addin/client-interfaces/addin-confirm-button-style.ts
@@ -1,0 +1,14 @@
+/**
+ * Defines the style of a confirm dialog button.
+ */
+export enum AddinConfirmButtonStyle {
+  // The button represents a "secondary" action (default)
+  Default = 0,
+
+  // The button represents a "primary" action
+  Primary = 1,
+
+  // The button represents a "link" action, often used
+  // to trigger an action to cancel/close a confirm dialog
+  Link = 2
+}